First Real Commit
I have a board
I can place tiles on the board
They are random pieces and random colours and they just overwrite one
another. There's no checking if they'll fit yet

https://www.youtube.com/watch?v=ODLzYI4d-J8&ab_channel=Zigurous
Found a tutorial for how to make tetris so I'm using some of that. I had
already started before I found the tutorial so some stuff is different. I
wasn't sure what a 'piece' is or how to rotate them. This implementation
uses some matrix math to rotate them and pieces are stored in PieceData as
x,y offsets from a center point.
My game is only kinda tetris so only some of the tutorial is useful. Like
you can place pieces where you want so I don't need the gravity or wall
kicks stuff. And you're going to hold the current piece above the field so
it might be messy to use it with tile maps. I think it'll be easier to
have the held tile be a different object instead of constantly swapping
the tile map to show tiles. Normal tetris doesn't have overlapping like
you get when you hold a piece over the board before you place it.