Preview
Can Place tiles and it checks if they are allowed to be placed there,
Won't let you place tiles over other tiles or out of bounds.

Preview
Shows what tile you will place and what colour it will be.
Attached to the mouse, not to the grid at the moment.
I feel like I did some hacky stuff to make it work that I might need to
fix. Like PreviewShape having a ref to the playerInput for the mouse pos
and TileGrid having a ref to Preview to send the next piece. Can I do
those better? Event for OnNextTimeGenerated? Might have to rethink this
when I make it snap to grid anyway.
Also feel like I'm using shape and tetromino interchangebly and maybe I
shouldn't be? The link between shape and colour might not be consistent in
different classes. One stores them as a single object and others are
separate?