Match 3
When you place a tile, check the surroundings to see if it's a match 3.
For each tile in the shape, check all neighbours for surrounding blocks.
Check if the neighbouring shapes are the same colour and not yourself.
Then do the same for any shapes that are the same colour to see if they
have any neighbours that match.

Needed to make the shape struct comparable and unique so I can do
.equals() and GetHashCode() for use with a hashset<Shape>.
The hash formula might not be good enough, but it's probably pretty rare
to have overlap. Maybe O tetros on the left and bottom edge overlap their
hashes? bc the hash is tetro*x^y and 1^2 == 1^6.

Nothing happens when you get a match 3 yet. But I can detect them.

Tiles and their sense of self
Tiles know what shape they are a part of and add themselves to that shape
So from a tile, you can get its shape and all other members
I'm gonna have some bugs with this when shapes start getting cut apart by
tetrises. The AddMember and RemoveMember aren't exactly robust or safe
atm.