Rotation and Horizontal Lines
You can right click to rotate the piece clockwise
Copied the rotation logic from the tutorial
And it seems like it worked first try
Shapes only use their cells when placed and when I get a new shape, it
just creates a new shape. So it's fine to just edit the shapes cells. I
was worried this might not exactly be possible, but it totally is. Made it
much simpler.

Horizontal Lines
Now checks for horizontal lines in addition to verical.
Made the code for the E/W checking a bit better. Like updating the current
left/rightmost tile as I go instead of at the end for no reason.

Bugs
Fixed a bug in the shape detection
a neighbour shape can be added to the stack multiple times. So now when it
pops, it keeps popping until it finds a shape it hasn't evaluated yet.

Line base case
When you check for a line, you add the current tile bc it's guarenteed to
be in the line. Except I wasn't adding the current tile, I was always
adding the center tile of the shape instead of the current tile. Which mostly works. The center tile
is the most likely tile to be in a line so edge cases are rather rare.